refactor: isolate fund results route state wrappers
Terminal page states and the scroll fade wrapper are route-adjacent UI shells, so moving them out reduces page size without touching data semantics.

Constraint: Keep rendered copy, route fallback behavior, and wouter navigation unchanged.

Confidence: medium

Scope-risk: narrow

Tested: npx vitest client fund-model-results page test

Not-tested: Full client suite
