+++ client/src/pages/fund-model-results.tsx
@@ -11,15 +11,8 @@
  */
 
 import React, { useRef, useState, useEffect, useCallback } from 'react';
-import { Link, useRoute, useLocation } from 'wouter';
-import {
-  AlertCircle,
-  ArrowLeft,
-  ChevronDown,
-  ChevronRight,
-  History,
-  RefreshCw,
-} from 'lucide-react';
+import { Link, useRoute } from 'wouter';
+import { AlertCircle, ChevronDown, ChevronRight, History, RefreshCw } from 'lucide-react';
 import { Alert, AlertTitle, AlertDescription } from '@/components/ui/alert';
 import { Button } from '@/components/ui/button';
 import { Badge } from '@/components/ui/badge';
@@ -60,6 +53,8 @@ import {
   sectionBackedEvidence,
   sectionEvidence,
 } from './fund-model-results/evidence';
+import { FadeInSection } from './fund-model-results/FadeInSection';
+import { ErrorState, LatestErrorState, LoadingState } from './fund-model-results/states';
 import {
   capitalize,
   diagnosticAlertClasses,
@@ -97,54 +92,6 @@ function isActiveCalculationStatus(status: LifecycleStatus) {
 
 function isTerminalCalculationStatus(status: LifecycleStatus) {
   return status === 'ready' || status === 'failed' || status === 'not_requested';
-}
-
-// ============================================================================
-// HOOKS
-// ============================================================================
-
-/**
- * Fade-in effect triggered once when the element enters the viewport.
- * Uses IntersectionObserver with a 10% visibility threshold.
- */
-function useFadeInOnScroll() {
-  const ref = useRef<HTMLDivElement>(null);
-  const [isVisible, setIsVisible] = useState(false);
-
-  useEffect(() => {
-    const el = ref.current;
-    if (!el) return;
-    const observer = new IntersectionObserver(
-      ([entry]) => {
-        if (entry?.isIntersecting) {
-          setIsVisible(true);
-          observer.disconnect();
-        }
-      },
-      { threshold: 0.1 }
-    );
-    observer.observe(el);
-    return () => observer.disconnect();
-  }, []);
-
-  return { ref, isVisible };
-}
-
-/** Wrapper div that fades in when scrolled into view */
-function FadeInSection({ children }: { children: React.ReactNode }) {
-  const { ref, isVisible } = useFadeInOnScroll();
-
-  return (
-    <div
-      ref={ref}
-      className={cn(
-        'transition-all duration-700 ease-out',
-        isVisible ? 'opacity-100 translate-y-0' : 'opacity-0 translate-y-4'
-      )}
-    >
-      {children}
-    </div>
-  );
 }
 
 // ============================================================================
@@ -1512,67 +1459,6 @@ function SectionRenderer({
         {statusLabel ? `${statusLabel}: ` : ''}
         {copy}
       </p>
-    </div>
-  );
-}
-
-// ============================================================================
-// STATE COMPONENTS
-// ============================================================================
-
-function LoadingState() {
-  return (
-    <div className="max-w-6xl mx-auto px-6 py-24 text-center" role="status">
-      <div className="animate-pulse space-y-4">
-        <div className="h-8 bg-beige-100 rounded w-1/3 mx-auto" />
-        <div className="h-4 bg-beige-100 rounded w-1/2 mx-auto" />
-        <div className="h-32 bg-beige-100 rounded" />
-        <div className="h-32 bg-beige-100 rounded" />
-      </div>
-    </div>
-  );
-}
-
-function LatestErrorState() {
-  const [, navigate] = useLocation();
-  return (
-    <div className="max-w-xl mx-auto px-8 py-24 text-center">
-      <Alert className="mb-8 border-beige-200">
-        <AlertCircle className="h-5 w-5 text-charcoal-400" />
-        <AlertTitle>Invalid results route</AlertTitle>
-        <AlertDescription className="font-poppins text-charcoal-500">
-          No fund ID specified. Please complete the modeling wizard to view results.
-        </AlertDescription>
-      </Alert>
-      <Button
-        variant="outline"
-        className="border-charcoal-300 text-charcoal hover:bg-charcoal-50"
-        onClick={() => navigate('/fund-setup')}
-      >
-        <ArrowLeft className="h-4 w-4 mr-2" />
-        Go to Fund Setup
-      </Button>
-    </div>
-  );
-}
-
-function ErrorState({ message }: { message: string }) {
-  const [, navigate] = useLocation();
-  return (
-    <div className="max-w-xl mx-auto px-8 py-24 text-center">
-      <Alert className="mb-8 border-beige-200">
-        <AlertCircle className="h-5 w-5 text-charcoal-400" />
-        <AlertTitle>Error loading results</AlertTitle>
-        <AlertDescription className="font-poppins text-charcoal-500">{message}</AlertDescription>
-      </Alert>
-      <Button
-        variant="outline"
-        className="border-charcoal-300 text-charcoal hover:bg-charcoal-50"
-        onClick={() => navigate('/fund-setup')}
-      >
-        <ArrowLeft className="h-4 w-4 mr-2" />
-        Back to Fund Setup
-      </Button>
     </div>
   );
 }
