+++ client/src/pages/fund-model-results/FadeInSection.tsx
@@ -1,0 +1,54 @@
+/**
+ * IntersectionObserver fade-in wrapper for the fund model results route.
+ *
+ * Extracted unchanged from client/src/pages/fund-model-results.tsx.
+ *
+ * @module client/pages/fund-model-results/FadeInSection
+ */
+
+import React, { useEffect, useRef, useState } from 'react';
+import { cn } from '@/lib/utils';
+
+/**
+ * Fade-in effect triggered once when the element enters the viewport.
+ * Uses IntersectionObserver with a 10% visibility threshold.
+ */
+function useFadeInOnScroll() {
+  const ref = useRef<HTMLDivElement>(null);
+  const [isVisible, setIsVisible] = useState(false);
+
+  useEffect(() => {
+    const el = ref.current;
+    if (!el) return;
+    const observer = new IntersectionObserver(
+      ([entry]) => {
+        if (entry?.isIntersecting) {
+          setIsVisible(true);
+          observer.disconnect();
+        }
+      },
+      { threshold: 0.1 }
+    );
+    observer.observe(el);
+    return () => observer.disconnect();
+  }, []);
+
+  return { ref, isVisible };
+}
+
+/** Wrapper div that fades in when scrolled into view */
+export function FadeInSection({ children }: { children: React.ReactNode }) {
+  const { ref, isVisible } = useFadeInOnScroll();
+
+  return (
+    <div
+      ref={ref}
+      className={cn(
+        'transition-all duration-700 ease-out',
+        isVisible ? 'opacity-100 translate-y-0' : 'opacity-0 translate-y-4'
+      )}
+    >
+      {children}
+    </div>
+  );
+}
